+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Mock gyro and sensor fusion
 
 ### Changed
 - Switching Projects to original build system for Xcode 10 (workaround for CocoaPod issue)